Change packet decoder API
