+++ crates/valence/src/client.rs
@@ -13,7 +13,8 @@ use rand::Rng;
 use tracing::warn;
 use valence_protocol::block_pos::BlockPos;
 use valence_protocol::byte_angle::ByteAngle;
-use valence_protocol::codec::{PacketDecoder, PacketEncoder};
+use valence_protocol::decoder::PacketDecoder;
+use valence_protocol::encoder::PacketEncoder;
 use valence_protocol::ident::Ident;
 use valence_protocol::item::ItemStack;
 use valence_protocol::packet::s2c::play::game_state_change::GameEventKind;
