+++ crates/valence/src/packet.rs
@@ -1,7 +1,7 @@
 use std::io::Write;
 
 use tracing::warn;
-use valence_protocol::codec::{encode_packet, encode_packet_compressed, PacketEncoder};
+use valence_protocol::encoder::{encode_packet, encode_packet_compressed, PacketEncoder};
 use valence_protocol::Packet;
 
 /// Types that can have packets written to them.
