+++ crates/valence/src/server/connection.rs
@@ -9,7 +9,8 @@ use tokio::sync::OwnedSemaphorePermit;
 use tokio::task::JoinHandle;
 use tokio::time::timeout;
 use tracing::debug;
-use valence_protocol::codec::{PacketDecoder, PacketEncoder};
+use valence_protocol::decoder::{decode_packet, PacketDecoder};
+use valence_protocol::encoder::PacketEncoder;
 use valence_protocol::Packet;
 
 use crate::client::{ClientBundle, ClientConnection};
@@ -23,6 +24,7 @@ pub(super) struct InitialConnection<R, W> {
     writer: W,
     enc: PacketEncoder,
     dec: PacketDecoder,
+    frame: BytesMut,
     timeout: Duration,
     permit: OwnedSemaphorePermit,
 }
@@ -47,6 +49,7 @@ where
             writer,
             enc,
             dec,
+            frame: BytesMut::new(),
             timeout,
             permit,
         }
@@ -67,30 +70,11 @@ where
         P: Packet<'a>,
     {
         timeout(self.timeout, async {
-            while !self.dec.has_next_packet()? {
-                self.dec.reserve(READ_BUF_SIZE);
-                let mut buf = self.dec.take_capacity();
-
-                if self.reader.read_buf(&mut buf).await? == 0 {
-                    return Err(io::Error::from(ErrorKind::UnexpectedEof).into());
-                }
-
-                // This should always be an O(1) unsplit because we reserved space earlier and
-                // the call to `read_buf` shouldn't have grown the allocation.
-                self.dec.queue_bytes(buf);
-            }
-
-            Ok(self
-                .dec
-                .try_next_packet()?
-                .expect("decoder said it had another packet"))
-
-            // The following is what I want to write but can't due to borrow
-            // checker errors I don't understand.
-            /*
             loop {
-                if let Some(pkt) = self.dec.try_next_packet()? {
-                    return Ok(pkt);
+                if let Some(frame) = self.dec.try_next_packet()? {
+                    self.frame = frame;
+
+                    return decode_packet(&self.frame);
                 }
 
                 self.dec.reserve(READ_BUF_SIZE);
@@ -104,7 +88,6 @@ where
                 // the call to `read_buf` shouldn't have grown the allocation.
                 self.dec.queue_bytes(buf);
             }
-            */
         })
         .await?
     }
@@ -112,7 +95,7 @@ where
     #[allow(dead_code)]
     pub fn set_compression(&mut self, threshold: Option<u32>) {
         self.enc.set_compression(threshold);
-        self.dec.set_compression(threshold.is_some());
+        self.dec.set_compression(threshold);
     }
 
     pub fn enable_encryption(&mut self, key: &[u8; 16]) {
