+++ crates/valence_protocol/src/decoder.rs
@@ -1,0 +1,257 @@
+#[cfg(feature = "encryption")]
+use aes::cipher::{AsyncStreamCipher, NewCipher};
+use anyhow::{bail, ensure};
+use bytes::{Buf, BufMut, BytesMut};
+
+use crate::var_int::{VarInt, VarIntDecodeError};
+use crate::{Encode, Packet, Result, MAX_PACKET_SIZE};
+
+/// The AES block cipher with a 128 bit key, using the CFB-8 mode of
+/// operation.
+#[cfg(feature = "encryption")]
+type Cipher = cfb8::Cfb8<aes::Aes128>;
+
+#[cfg(feature = "compression")]
+pub fn encode_packet_compressed<'a, P>(
+    buf: &mut Vec<u8>,
+    pkt: &P,
+    threshold: u32,
+    scratch: &mut Vec<u8>,
+) -> Result<()>
+where
+    P: Packet<'a>,
+{
+    use std::io::Read;
+
+    use flate2::bufread::ZlibEncoder;
+    use flate2::Compression;
+
+    let start_len = buf.len();
+
+    pkt.encode_packet(&mut *buf)?;
+
+    let data_len = buf.len() - start_len;
+
+    if data_len > threshold as usize {
+        let mut z = ZlibEncoder::new(&buf[start_len..], Compression::new(4));
+
+        scratch.clear();
+
+        let data_len_size = VarInt(data_len as i32).written_size();
+
+        let packet_len = data_len_size + z.read_to_end(scratch)?;
+
+        ensure!(
+            packet_len <= MAX_PACKET_SIZE as usize,
+            "packet exceeds maximum length"
+        );
+
+        drop(z);
+
+        buf.truncate(start_len);
+
+        VarInt(packet_len as i32).encode(&mut *buf)?;
+        VarInt(data_len as i32).encode(&mut *buf)?;
+        buf.extend_from_slice(scratch);
+    } else {
+        let data_len_size = 1;
+        let packet_len = data_len_size + data_len;
+
+        ensure!(
+            packet_len <= MAX_PACKET_SIZE as usize,
+            "packet exceeds maximum length"
+        );
+
+        let packet_len_size = VarInt(packet_len as i32).written_size();
+
+        let data_prefix_len = packet_len_size + data_len_size;
+
+        buf.put_bytes(0, data_prefix_len);
+        buf.copy_within(start_len..start_len + data_len, start_len + data_prefix_len);
+
+        let mut front = &mut buf[start_len..];
+
+        VarInt(packet_len as i32).encode(&mut front)?;
+        // Zero for no compression on this packet.
+        VarInt(0).encode(front)?;
+    }
+
+    Ok(())
+}
+
+#[derive(Default)]
+pub struct PacketDecoder {
+    buf: BytesMut,
+    #[cfg(feature = "compression")]
+    decompress_buf: BytesMut,
+    #[cfg(feature = "compression")]
+    compression_threshold: Option<u32>,
+    #[cfg(feature = "encryption")]
+    cipher: Option<Cipher>,
+}
+
+impl PacketDecoder {
+    pub fn new() -> Self {
+        Self::default()
+    }
+
+    pub fn try_next_packet(&mut self) -> Result<Option<BytesMut>> {
+        let mut r = &self.buf[..];
+
+        let packet_len = match VarInt::decode_partial(&mut r) {
+            Ok(len) => len,
+            Err(VarIntDecodeError::Incomplete) => return Ok(None),
+            Err(VarIntDecodeError::TooLarge) => bail!("malformed packet length VarInt"),
+        };
+
+        ensure!(
+            (0..=MAX_PACKET_SIZE).contains(&packet_len),
+            "packet length of {packet_len} is out of bounds"
+        );
+
+        if r.len() < packet_len as usize {
+            // Not enough data arrived yet.
+            return Ok(None);
+        }
+
+        let packet_len_len = VarInt(packet_len).written_size();
+
+        #[cfg(feature = "compression")]
+        if let Some(threshold) = self.compression_threshold {
+            use std::io::Write;
+
+            use flate2::write::ZlibDecoder;
+
+            use crate::Decode;
+
+            r = &r[..packet_len as usize];
+
+            let data_len = VarInt::decode(&mut r)?.0;
+
+            ensure!(
+                (0..MAX_PACKET_SIZE).contains(&data_len),
+                "decompressed packet length of {data_len} is out of bounds"
+            );
+
+            // Is this packet compressed?
+            if data_len > 0 {
+                ensure!(
+                    data_len as u32 > threshold,
+                    "decompressed packet length of {data_len} is <= the compression threshold of \
+                     {threshold}"
+                );
+
+                debug_assert!(self.decompress_buf.is_empty());
+
+                self.decompress_buf.put_bytes(0, data_len as usize);
+
+                // TODO: use libdeflater or zune-inflate?
+                let mut z = ZlibDecoder::new(&mut self.decompress_buf[..]);
+
+                z.write_all(r)?;
+
+                ensure!(
+                    z.finish()?.is_empty(),
+                    "decompressed packet length is shorter than expected"
+                );
+
+                let total_packet_len = VarInt(packet_len).written_size() + packet_len as usize;
+
+                self.buf.advance(total_packet_len);
+
+                return Ok(Some(self.decompress_buf.split()));
+            } else {
+                debug_assert_eq!(data_len, 0);
+
+                ensure!(
+                    r.len() <= threshold as usize,
+                    "uncompressed packet length of {} exceeds compression threshold of {}",
+                    r.len(),
+                    threshold
+                );
+
+                let remaining_len = r.len();
+
+                self.buf.advance(packet_len_len + 1);
+                return Ok(Some(self.buf.split_to(remaining_len)));
+            }
+        }
+
+        self.buf.advance(packet_len_len);
+        Ok(Some(self.buf.split_to(packet_len as usize)))
+    }
+
+    #[cfg(feature = "compression")]
+    pub fn compression(&self) -> Option<u32> {
+        self.compression_threshold
+    }
+
+    #[cfg(feature = "compression")]
+    pub fn set_compression(&mut self, threshold: Option<u32>) {
+        self.compression_threshold = threshold;
+    }
+
+    #[cfg(feature = "encryption")]
+    pub fn enable_encryption(&mut self, key: &[u8; 16]) {
+        assert!(self.cipher.is_none(), "encryption is already enabled");
+
+        let mut cipher = Cipher::new(key.into(), key.into());
+
+        // Don't forget to decrypt the data we already have.
+        cipher.decrypt(&mut self.buf);
+
+        self.cipher = Some(cipher);
+    }
+
+    pub fn queue_bytes(&mut self, mut bytes: BytesMut) {
+        #![allow(unused_mut)]
+
+        #[cfg(feature = "encryption")]
+        if let Some(cipher) = &mut self.cipher {
+            cipher.decrypt(&mut bytes);
+        }
+
+        self.buf.unsplit(bytes);
+    }
+
+    pub fn queue_slice(&mut self, bytes: &[u8]) {
+        #[cfg(feature = "encryption")]
+        let len = self.buf.len();
+
+        self.buf.extend_from_slice(bytes);
+
+        #[cfg(feature = "encryption")]
+        if let Some(cipher) = &mut self.cipher {
+            cipher.decrypt(&mut self.buf[len..]);
+        }
+    }
+
+    #[deprecated]
+    pub fn queued_bytes(&self) -> &[u8] {
+        // TODO: skip prepared packet?
+        self.buf.as_ref()
+    }
+
+    pub fn take_capacity(&mut self) -> BytesMut {
+        self.buf.split_off(self.buf.len())
+    }
+
+    pub fn reserve(&mut self, additional: usize) {
+        self.buf.reserve(additional);
+    }
+}
+
+/// Decodes a (packet ID + data) packet frame. An error is returned if the input
+/// is not read to the end.
+pub fn decode_packet<'a, P: Packet<'a>>(mut bytes: &'a [u8]) -> anyhow::Result<P> {
+    let pkt = P::decode_packet(&mut bytes)?;
+
+    ensure!(
+        bytes.is_empty(),
+        "missed {} bytes while decoding {}",
+        bytes.len(),
+        pkt.packet_name()
+    );
+
+    Ok(pkt)
+}
