+++ crates/valence_protocol/src/var_int.rs
@@ -18,7 +18,7 @@ impl VarInt {
 
     /// Returns the exact number of bytes this varint will write when
     /// [`Encode::encode`] is called, assuming no error occurs.
-    pub fn written_size(self) -> usize {
+    pub const fn written_size(self) -> usize {
         match self.0 {
             0 => 1,
             n => (31 - n.leading_zeros() as usize) / 7 + 1,
